Move FrameOptions and related function to module itself
